Storing component state in context.
React context is experimental and discouraged.  However, it’s safe if you use it simply as dependency injection.  It seems the best solution for this project, since I want people to be able to combine the components in any combination they require so long as it is surrounded by the CarouselProvider component.  This is a savepoint commit.