Uglify breaks the es module
I tried the experimental uglify-es but it seems to rename all the JSX tag names to random letters.  So, I’ll wait and revisit in the future.  For the most part, if you are using an ES module, it’s going to get transpiled and minified by the consumer anyhow.