bugfix carousel sometimes became misaligned
window.requestAnimation frame would sometimes fire after touchEnd if you moved the carousel back and forth a bunch of times.  That’s because the rendering caused the browser to find an animation frame that occured after the touch ended.  I needed to add cancelAnimationFrame calls in the appropriate places.

Also added a “polyfill” of sorts for requestAnimationFrame.

Also the requestAnimationFrame timer was never being cleaned up on unmount, that causes a memory leak under some circumstances.  That’s fixed.