bugfix vertical scrolling issue when scrolling on carousel
Code intended only for a vertical carousel was accidentally being applied to horizontal carousels.  When a user touchMoves vertically on a horizontal carousel, the page should scroll.  It was scrolling in a jerky motion.  Very clunky.  Because code for vertical scrolling was cutting the page off and preventing a scroll.