added option for different image on image zoom
we had complaints that the un-zoomed carousel images looked fuzzy.  It’s because we were taking the larger, “zoomed” image and squishing it into a smaller space.  That’s ok most of the time, but we learned from an adobe consultant, that some squishing is ok, a lot can result in fuzzy images on webkit browsers.  Firefox handles it ok.  The fuzz is subtle, but the creatives and marketing folks noticed the fuzz and can’t stop seeing it.