Feature/passing though events to slider (#145)
* passing through events on Slider Component

A goal of the project is to pass through all props when possible.  We forgot to pass through event handlers on the slider component.

* updated packages to silence rollup warnings

Needed to update some rollup plugins after recent update to rollup.  These plugins were using deprecated rollup hooks.

* passing through props to slider Tray tag

There was a request to pass through props to the Tray tag in slider.

* Update Slider.jsx (#146)