Add two props to DotGroup
### disableActiveDots
#### Current behaviour
When a dot is shown as selected, it is disabled. So when two slides are visibles, one is fully visible and the other one is partially visible, we can't click on the dot corresponding to the partially visible slide. 
#### Improved behaviour
This new prop does not change the default behaviour (so no breaking change), but it adds the option to let the dots enabled whatever, even if the slide is already visible.

### showAsSelectedForCurrentSlideOnly
#### Current behaviour
For now, the dots are shown as selected is the slide is visible, so there might be several dots shown as selected. We can also want to show the dot as selected for the current slide only.
#### Improved behaviour
This new prop does not change the default behaviour (so no breaking change), but it adds the option to show the dot as selected for the current slide only.