build(packaging): various updates to build process and security
* fix the demo page

The demo page was broken because the last deploy was done with the wrong version of node and npm.  Added a .nvmrc file and updated the path for the css file.  Also updated README documentation.

* Adding package-lock.json to repo

Supposedly this was removed to make it easier for contributors to contribute, but after researching the matter the best practice is to always commit a lockfile so that everyone is working from the same source of truth.

* security update for semantic-release

Fixing github security warning

* gitignore the output directory for dev server

* fix css not showing up in local dev server

* these files are now gitignored

* Force people to use the correct node version

* Update ci.yml

set node version to 10.16.1